requirements.txt
transformers
torch
gradio
opencv-python-headless
numpy
pillow